requirements.txt
python-telegram-bot==20.3
openai==1.3.8
requests==2.31.0
python-dotenv==1.0.0
pytrends==4.9.2
jinja2==3.1.2
Pillow==10.1.0
httpx==0.24.1
